feature(categories-api): create updateAnswer handler ## Implementation of updating only changed fields at the Go level
In this implementation, we only update the fields that have changed at the Go code level. This makes it easier to test and maintain the code in the future.

In this implementation, we use the `updateAnswerParams` structure to deserialize the JSON request. The `IsCorrect` field is declared as a pointer to `bool` so that we can determine if the value was specified in the request or not.

In the `updateAnswer` handler, we get the current answer value from the database and then check each field of the request. If the `Text` field is not specified, we save the current value from the database. If the `IsCorrect` field is `nil`, we also save the current value from the database.

We then create an `arg` object of type `db.UpdateAnswerParams` which is used to call the `UpdateAnswer` method on the data store.

In this way, we update only changed fields at the level of the Go code, which allows for more efficient processing of requests and makes it easier to test and maintain the code.

> **Note:** A SQL-level implementation using the `COALESCE` function is also possible, but in this case we chose a Go code-level implementation because it makes it easier to test and maintain the code in the future.